Revert `->getPrimary()` to `->id`
Few changes were needed for `findBy` blocks. Using `->findBy(...)->id`
generates notice: "Trying to get property 'id' of non-object". With
`->findBy(...)->getPrimary()` this wasn't problem. Throwing exception
(instead of notice), this is error when creating elements.

Reason of change: To be consistent with the rest of application.
Maybe we'll refactor this one day.

remp/crm#449