Fixed issue with triggered before events when regular trigger event was triggered
When regular trigger was dispatched (e.g. `subscription_ends`) and there is
a scenario with before event and same name (e.g. before `subscription_ends`),
dispatcher triggered all before events as well due to the missing condition.

This caused some notification emails to be sent in not very appropriate time.