Allow scenario condition to have empty parameter
The recently added validation forced each parameter
to be used, which might not always be the intended
behavior. At least one parameter of criteria should
always be specified, but we can't enforce all of them
to be entered.

This could be possible flagged by criteria themselves
in the future, but for now, "at least one param"
is our evaluation of choice.

remp/crm#3339